content: use semaphore instead of backoff and retry
Both content backend plugin and OpenWriter helper use backoff and retry
to hold reference for importing blob data. It can save network/IO
bandwidth when concurrent download request on the same blob data.

The maximum retry interval can be 2s. If there are several concurrent
requests for the same image, it will take long time to finish all the
image pulling jobs. And it will be worse if the image has many more
layers.

In order to improve the performance, use a weighted semaphore instead
of backoff and retry. If the content.Writer has been hold, we should
park the fetching-the-same-blob goroutines into waiting state. When the
writer has been released, the one of waiting goroutines will be notified
and hold the writer. It can reduce useless sleep.

And this patch makes MaxConcurrentDownloads tunable for CRI-plugin.
The MaxConcurrentDownloads is sharing between several image pulling
requests. It is aligned with moby engine. It is good to keep the value
small to reduce impacts for running containers when pulling images.

Fixes: #4937

Signed-off-by: Wei Fu <fuweid89@gmail.com>